(200619)Ahna_IndoorPipline, Bose Greeting

+++ Client/CorruptLab_Client/CorruptLab/Object_Player.h
@@ -44,8 +44,8 @@ protected:
 
 //GameContents
 	int					   m_iState;
-	int				       m_iMaxHP ;
-	int				       m_iCurrentHP ;
+	float				   m_iMaxHP ;
+	float			       m_iCurrentHP ;
 	int				       m_iAtt ;
 
 ///
@@ -54,7 +54,7 @@ public:
 //GameContents
 	void GetDamaage(int iDamage); 
 	void GetFullHP() { m_iCurrentHP = m_iMaxHP; }
-	int* GetPlayerHPPointer() { return &m_iCurrentHP; }
+	float* GetPlayerHPPointer() { return &m_iCurrentHP; }
 ///
